backend/requirements.txt
annotated-types==0.6.0
anyio==4.2.0
asgiref==3.7.2
backoff==2.2.1
bcrypt==4.1.2
cachetools==5.3.2
certifi==2023.11.17
charset-normalizer==3.3.2
chromadb-client==0.4.19.dev0
click==8.1.7
coloredlogs==15.0.1
Deprecated==1.2.14
distro==1.9.0
exceptiongroup==1.2.0
fastapi==0.108.0
filelock==3.13.1
flatbuffers==23.5.26
fsspec==2023.12.2
google-auth==2.25.2
googleapis-common-protos==1.62.0
grpcio==1.60.0
h11==0.14.0
httpcore==1.0.2
httptools==0.6.1
httpx==0.26.0
huggingface-hub==0.20.1
humanfriendly==10.0
idna==3.6
importlib-metadata==6.11.0
importlib-resources==6.1.1
kubernetes==28.1.0
mangum==0.17.0
mmh3==4.0.1
monotonic==1.6
mpmath==1.3.0
numpy==1.26.2
oauthlib==3.2.2
onnxruntime==1.16.3
openai==1.6.1
opentelemetry-api==1.22.0
opentelemetry-exporter-otlp-proto-common==1.22.0
opentelemetry-exporter-otlp-proto-grpc==1.22.0
opentelemetry-instrumentation==0.43b0
opentelemetry-instrumentation-asgi==0.43b0
opentelemetry-instrumentation-fastapi==0.43b0
opentelemetry-proto==1.22.0
opentelemetry-sdk==1.22.0
opentelemetry-semantic-conventions==0.43b0
opentelemetry-util-http==0.43b0
overrides==7.4.0
packaging==23.2
posthog==3.1.0
protobuf==4.25.1
pulsar-client==3.3.0
pyasn1==0.5.1
pyasn1-modules==0.3.0
pydantic==1.10.13
pydantic_core==2.14.6
PyPika==0.48.9
python-dateutil==2.8.2
python-dotenv==1.0.0
PyYAML==6.0.1
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
sniffio==1.3.0
starlette==0.32.0.post1
sympy==1.12
tenacity==8.2.3
tokenizers==0.15.0
tqdm==4.66.1
typer==0.9.0
typing_extensions==4.9.0
urllib3==1.26.18
uvicorn==0.25.0
uvloop==0.19.0
watchfiles==0.21.0
websocket-client==1.7.0
websockets==12.0
wrapt==1.16.0
zipp==3.17.0
